pyproject.toml
[tool.poetry]
name = "panso"
version = "0.1.0"
description = "Panso.se"
authors = ["Joakim Hellsén <tlovinator@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
django = "^4.2.7"
django-simple-history = "^3.4.0"
sitemap-parser = { git = "https://github.com/TheLovinator1/sitemap-parser.git" }
typer = { extras = ["all"], version = "^0.9.0" }
django-ninja = "^1.1.0"
redis = { extras = ["hiredis"], version = "^5.0.1" }
psycopg = { extras = ["binary"], version = "^3.1.13" }
tenacity = "^8.2.3"
orjson = "^3.9.10"
whitenoise = { extras = ["brotli"], version = "^6.6.0" }
selectolax = "^0.3.17"
hishel = "^0.0.19"
httpx = { extras = ["brotli", "http2"], version = "^0.25.2" }
playwright = "^1.40.0"
dateparser = "^1.2.0"
gunicorn = "^21.2.0"
tablib = { extras = ["all"], version = "^3.5.0" }
django-filter = "^23.5"
celery = { extras = ["redis"], version = "^5.3.6" }
django-celery-results = "^2.5.1"
flower = "^2.0.1"
django-debug-toolbar = "^4.2.0"
django-cacheops = "^7.0.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pytest = "^7.4.3"
djlint = "^1.34.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.ruff]
exclude = ["migrations"]
fix = true
unsafe-fixes = true
preview = true
select = ["ALL"]
ignore = [
  "CPY001", # Missing copyright notice at top of file
  "ERA001", # Found commented-out code
  "FIX002", # Line contains TODO
  "DJ001",  # Allow TextField to be nullable
  # TODO: #25 We should probably remove DJ001 and every model that has null=True
  # https://github.com/TheLovinator1/panso.se/issues/25
]
line-length = 120

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
"**/tests.py" = [
  "S101",    # Allow asserts
  "ARG",     # Allow unused arguments
  "FBT",     # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
  "PLR2004", # Allow "assert response.status_code == 200" when testing views
  "D102",    # Allow missing docstrings in tests
  "PLR6301", # Checks for the presence of unused self parameter in methods definitions.
]
